--- src/vscode-avalonia/src/util/AppConstants.ts
@@ -1,13 +1,0 @@
-export class AppConstants {
-	static readonly previewerParamState = "previewerParams";
-	static readonly previewProcessCommandId = "avalonia.previewProcess";
-	static readonly localhost = "http://127.0.0.1";
-
-	static readonly updateAssetsMessages: "updateAssetsMessage";
-	static readonly showPreivewMessage: "showPreviewMessage";
-
-	static readonly showPreviewToSideCommand = "avalonia.showPreviewToSide";
-	static readonly previewerAssetsCommand = "avalonia.createPreviewerAssets";
-
-	static readonly previewerPanelViewType = "avaloniaPreviewer";
-}
